Move QR/passcode views to separate files
